pipelines/press_releases/requirements.txt
pydantic
prettyprint
pytest==8.1.1
pytest-cov==5.0.0
pytest-html==4.1.1
pytest-metadata==3.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.27.1
schedule==1.2.1
Flask
Flask-SQLAlchemy
SQLAlchemy==2.0.25
SQLAlchemy-Utils==0.41.2
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==1.26.13
Werkzeug==2.3.5
marshmallow-sqlalchemy
psycopg2-binary==2.9.5
psycopg2==2.9.9
beautifulsoup4==4.12.3
html5lib==1.1
pandas==2.2.3
fastapi[standard]

# langchain
# langchain_community
# langchain_core
# langchain-ollama

elasticsearch
python-pushover